Some overall quality improvements
After game has ended by checkmate , engine is reset and loop resets to ready , simply rematch or look for new opponent and press enter to restart the engine. In the case of any other game end condition there is a 30 second timer and if no moves are made engine is reset and loop resets.